replace proteus with rsocket-rpc
update dependencies

gradle 5
