+++ .gitignore
@@ -242,3 +242,5 @@ gradle-app.setting
 # gradle/wrapper/gradle-wrapper.properties
 
 # End of https://www.gitignore.io/api/java,linux,macos,gradle,android,intellij,androidstudio
+
+generated/
