+++ grpc/src/main/java/io/netifi/thunderdome/grpc/nonblocking/rs/GrpcRequestReplyClient.java
@@ -10,6 +10,8 @@ import io.netty.channel.nio.NioEventLoopGroup;
 import io.netty.channel.socket.nio.NioSocketChannel;
 import io.netty.util.concurrent.DefaultThreadFactory;
 import org.HdrHistogram.Histogram;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import reactor.core.publisher.Flux;
 import reactor.core.scheduler.Schedulers;
 
@@ -17,6 +19,8 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.ThreadFactory;
 
 public class GrpcRequestReplyClient {
+  private static final Logger logger = LogManager.getLogger(GrpcRequestReplyClient.class);
+
   public static void main(String... args) throws Exception {
     int warmup = 1_000_000;
     int count = 1_000_000;
@@ -34,17 +38,17 @@ public class GrpcRequestReplyClient {
             .channelType(NioSocketChannel.class)
             .directExecutor()
             .flowControlWindow(NettyChannelBuilder.DEFAULT_FLOW_CONTROL_WINDOW)
-            .usePlaintext(true)
+            .usePlaintext()
             .build();
 
     SimpleServiceGrpc.SimpleServiceStub simpleService = SimpleServiceGrpc.newStub(managedChannel);
 
     Histogram histogram = new Histogram(3600000000000L, 3);
-    System.out.println("starting warmup...");
+    logger.info("starting warmup...");
     execute(count, simpleService, histogram, concurrency, executor);
-    System.out.println("warmup complete");
+    logger.info("warmup complete");
 
-    System.out.println("starting test - sending " + count);
+    logger.info("starting test - sending {}", count);
     histogram = new Histogram(3600000000000L, 3);
     long start = System.nanoTime();
     execute(count, simpleService, histogram, concurrency, executor);
@@ -52,9 +56,8 @@ public class GrpcRequestReplyClient {
     histogram.outputPercentileDistribution(System.out, 1000.0d);
     double completedMillis = (System.nanoTime() - start) / 1_000_000d;
     double rps = count / ((System.nanoTime() - start) / 1_000_000_000d);
-    System.out.println("test complete in " + completedMillis + "ms");
-    System.out.println("test rps " + rps);
-    System.out.println();
+    logger.info("test complete in {} ms", completedMillis);
+    logger.info("test rps {}", rps);
     System.exit(0);
   }
