+++ grpc/src/main/java/io/netifi/thunderdome/grpc/nonblocking/stock/GrpcRequestReplyClient.java
@@ -5,11 +5,15 @@ import io.grpc.ManagedChannelBuilder;
 import io.netifi.testing.protobuf.SimpleRequest;
 import io.netifi.testing.protobuf.SimpleServiceGrpc;
 import org.HdrHistogram.Histogram;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ForkJoinPool;
 
 public class GrpcRequestReplyClient {
+  private static final Logger logger = LogManager.getLogger(GrpcRequestReplyClient.class);
+
   public static void main(String... args) throws Exception {
     int warmup = 10_000;
     int count = 1_000_000;
@@ -18,14 +22,14 @@ public class GrpcRequestReplyClient {
     int port = Integer.getInteger("port", 8001);
 
     ManagedChannel managedChannel =
-        ManagedChannelBuilder.forAddress(host, port).usePlaintext(true).build();
+        ManagedChannelBuilder.forAddress(host, port).usePlaintext().build();
     SimpleServiceGrpc.SimpleServiceFutureStub simpleService =
         SimpleServiceGrpc.newFutureStub(managedChannel);
 
     CountDownLatch warmupLatch = new CountDownLatch(warmup);
 
     ForkJoinPool executorService = ForkJoinPool.commonPool();
-    System.out.println("starting warmup...");
+    logger.info("starting warmup...");
 
     for (int i = 0; i < warmup; i++) {
       SimpleRequest hello = SimpleRequest.newBuilder().setRequestMessage("hello").build();
@@ -33,9 +37,9 @@ public class GrpcRequestReplyClient {
     }
 
     warmupLatch.await();
-    System.out.println("warmup complete");
+    logger.info("warmup complete");
 
-    System.out.println("starting test - sending " + count);
+    logger.info("starting test - sending {}", count);
     CountDownLatch latch = new CountDownLatch(count);
     Histogram histogram = new Histogram(3600000000000L, 3);
     long start = System.nanoTime();
@@ -54,9 +58,8 @@ public class GrpcRequestReplyClient {
 
     double completedMillis = (System.nanoTime() - start) / 1_000_000d;
     double rps = count / ((System.nanoTime() - start) / 1_000_000_000d);
-    System.out.println("test complete in " + completedMillis + "ms");
-    System.out.println("test rps " + rps);
-    System.out.println();
+    logger.info("test complete in {} ms", completedMillis);
+    logger.info("test rps {}", rps);
     latch.await();
     managedChannel.shutdownNow();
   }
