+++ grpc/src/main/java/io/netifi/thunderdome/grpc/nonblocking/stock/GrpcServer.java
@@ -5,8 +5,13 @@ import io.grpc.stub.StreamObserver;
 import io.netifi.testing.protobuf.SimpleRequest;
 import io.netifi.testing.protobuf.SimpleResponse;
 import io.netifi.testing.protobuf.SimpleServiceGrpc;
+import io.netifi.thunderdome.grpc.nonblocking.rs.MultithreadedGrpcRequestReplyClient;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 public class GrpcServer {
+  private static final Logger logger = LogManager.getLogger(GrpcServer.class);
+
   static class DefaultService extends SimpleServiceGrpc.SimpleServiceImplBase {
     @Override
     public void requestReply(SimpleRequest request, StreamObserver<SimpleResponse> responseObserver) {
@@ -17,9 +22,8 @@ public class GrpcServer {
   }
   
   public static void main(String... args) throws Exception {
-    System.out.println("starting server");
-  
-  
+    logger.info("starting server");
+
     String host = System.getProperty("host", "127.0.0.1");
     int port = Integer.getInteger("port", 8001);
   
@@ -28,7 +32,7 @@ public class GrpcServer {
                                .build()
                                .start();
   
-    System.out.println("server started");
+    logger.info("server started");
     start.awaitTermination();
   }
 }
