+++ grpc/src/main/resources/log4j2.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Configuration status="warn">
+    <Appenders>
+        <Console name="STDOUT" target="SYSTEM_OUT">
+            <PatternLayout pattern="%d %p %c{1.} [%t] %m%n"/>
+        </Console>
+    </Appenders>
+    <Loggers>
+        <Root level="INFO" additivity="false">
+            <AppenderRef ref="STDOUT"/>
+        </Root>
+        <logger name="io.netifi" level="INFO" additivity="false">
+            <AppenderRef ref="STDOUT"/>
+        </logger>
+    </Loggers>
+</Configuration>
