--- proteus/build.gradle
@@ -1,60 +1,0 @@
-plugins {
-    id 'com.google.protobuf' version '0.8.3'
-    id 'java'
-    id 'idea'
-}
-
-sourceCompatibility = 1.8
-
-dependencies {
-    compile 'io.netifi.proteus:frames:0.4.0'
-    compile 'io.netifi.proteus:auth:0.4.0'
-    compile 'io.netifi.proteus:core:0.4.0'
-    compile 'org.hdrhistogram:HdrHistogram:2.1.10'
-    compile 'com.google.protobuf:protobuf-java:3.5.0'
-    compile 'io.rsocket:rsocket-transport-netty:0.9.16'
-}
-
-sourceSets {
-    main {
-        proto { srcDir 'src/main/proto' }
-    }
-
-    test {
-        proto { srcDir 'src/test/proto' }
-    }
-}
-
-protobuf {
-    generatedFilesBaseDir = "${projectDir}/src/generated"
-
-    protoc {
-        artifact = 'com.google.protobuf:protoc:3.5.0'
-    }
-
-    plugins {
-        proteus {
-            artifact = 'io.netifi.proteus:protobuf-rpc:0.4.0'
-        }
-        generateProtoTasks {
-            ofSourceSet('main')*.plugins {
-                proteus {}
-            }
-        }
-    }
-}
-
-idea {
-    module {
-        sourceDirs += file("src/main/proto")
-        sourceDirs += file("src/generated/main/java")
-        sourceDirs += file("src/generated/main/proteus")
-
-        generatedSourceDirs += file('src/generated/main/java')
-        generatedSourceDirs += file('src/generated/main/proteus')
-    }
-}
-
-clean {
-    delete 'src/generated'
-}
