+++ ratpack/build.gradle
@@ -1,18 +1,14 @@
 plugins {
-    id "io.ratpack.ratpack-java" version "1.5.0"
-    id 'java'
+    id "io.ratpack.ratpack-java" version "1.6.0"
     id 'idea'
 }
-
-sourceCompatibility = 1.8
 
 mainClassName = 'io.netifi.thunderdome.ratpack.RatpackTestServer'
 
 dependencies {
-    compile 'org.hdrhistogram:HdrHistogram:2.1.10'
-    compile 'io.projectreactor.ipc:reactor-netty:0.7.1.RELEASE'
-    compile 'org.slf4j:slf4j-simple:1.7.25'
+    compile "org.hdrhistogram:HdrHistogram:$hdrHistogramVersion"
+    compile "io.projectreactor.netty:reactor-netty:$reactorNettyVersion"
 
-    testCompile "junit:junit:4.12"
-    testCompile "org.mockito:mockito-core:1.9.5"
+    testCompile "junit:junit:$junitVersion"
+    testCompile "org.mockito:mockito-core:$mockitoVersion"
 }
