+++ ratpack/src/main/java/io/netifi/thunderdome/ratpack/RatpackRequestReply.java
@@ -5,16 +5,20 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import org.HdrHistogram.Histogram;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import reactor.ipc.netty.http.client.HttpClient;
+import reactor.netty.http.client.HttpClient;
+import reactor.netty.resources.ConnectionProvider;
 
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
+import java.io.InputStream;
 
 public class RatpackRequestReply {
+  private static final Logger logger = LogManager.getLogger(RatpackRequestReply.class);
 
-  public static void main(String... args) throws Exception {
+  public static void main(String... args) {
     int warmup = 1_000_000;
 
     int count = Integer.getInteger("count", 1_000_000);
@@ -25,107 +29,71 @@ public class RatpackRequestReply {
 
     ObjectMapper mapper = new ObjectMapper();
 
-    HttpClient client = HttpClient.create(host, port);
+    ConnectionProvider connectionProvider = ConnectionProvider
+        .fixed("test", concurrency);
+    HttpClient client = HttpClient
+        .create(connectionProvider)
+        .tcpConfiguration(c -> c.port(port).host(host));
 
-    System.out.println("starting warmup...");
+    logger.info("starting warmup...");
+
     Flux.range(0, warmup)
         .flatMap(
-            integer -> {
-              return client
-                  .post(
-                      "/",
-                      req -> {
-                        SimpleRequest request = new SimpleRequest();
-                        request.setMessage("hello");
-                        byte[] bytes = new byte[0];
-                        try {
-                          bytes = mapper.writeValueAsBytes(request);
-                        } catch (JsonProcessingException e) {
-                          e.printStackTrace();
-                        }
-                        return req.send(Mono.just(Unpooled.wrappedBuffer(bytes)));
-                      })
-                  .flatMap(
-                      res -> {
-                        StringBuilder builder = new StringBuilder();
-                        return res.receiveContent()
-                            .doOnNext(
-                                httpContent -> {
-                                  ByteBuf content = httpContent.content();
-                                  CharSequence charSequence =
-                                      content.readCharSequence(
-                                          content.readableBytes(), StandardCharsets.UTF_8);
-                                  builder.append(charSequence);
-                                })
-                            .then(
-                                Mono.fromRunnable(
-                                    () -> {
-                                      try {
-                                        mapper.readValue(builder.toString(), SimpleResponse.class);
-                                      } catch (IOException e) {
-                                        System.out.println(builder);
-                                        throw new RuntimeException(e);
-                                      }
-                                    }));
-                      });
-            })
+            integer -> client
+                .post()
+                .uri("/")
+                .send(newRequest(mapper))
+                .responseContent()
+                .aggregate()
+                .asInputStream()
+                .map(content -> toResponse(mapper, content)))
         .doOnError(Throwable::printStackTrace)
         .blockLast();
-    System.out.println("warmup complete");
+    logger.info("warmup complete");
 
     long start = System.nanoTime();
     Histogram histogram = new Histogram(3600000000000L, 3);
-    System.out.println("starting test - sending " + count);
+    logger.info("starting test - sending {}", count);
     Flux.range(0, count)
         .flatMap(
             integer -> {
               long s = System.nanoTime();
               return client
-                  .post(
-                      "/",
-                      req -> {
-                        SimpleRequest request = new SimpleRequest();
-                        request.setMessage("hello");
-                        byte[] bytes = new byte[0];
-                        try {
-                          bytes = mapper.writeValueAsBytes(request);
-                        } catch (JsonProcessingException e) {
-                          e.printStackTrace();
-                        }
-                        return req.send(Mono.just(Unpooled.wrappedBuffer(bytes)));
-                      })
-                  .flatMap(
-                      res -> {
-                        StringBuilder builder = new StringBuilder();
-                        return res.receiveContent()
-                            .doOnNext(
-                                httpContent -> {
-                                  ByteBuf content = httpContent.content();
-                                  CharSequence charSequence =
-                                      content.readCharSequence(
-                                          content.readableBytes(), StandardCharsets.UTF_8);
-                                  builder.append(charSequence);
-                                })
-                            .then(
-                                Mono.fromRunnable(
-                                    () -> {
-                                      try {
-                                        mapper.readValue(builder.toString(), SimpleResponse.class);
-                                      } catch (IOException e) {
-                                        System.out.println(builder);
-                                        throw new RuntimeException(e);
-                                      }
-                                      histogram.recordValue(System.nanoTime() - s);
-                                    }));
-                      });
+                  .post()
+                  .uri("/")
+                  .send(newRequest(mapper))
+                  .responseContent()
+                  .aggregate()
+                  .asInputStream()
+                  .map(content -> toResponse(mapper, content))
+                  .doOnNext(r -> histogram.recordValue(System.nanoTime() - s));
             },
             concurrency)
+        .doFinally(s -> connectionProvider.dispose())
         .blockLast();
     histogram.outputPercentileDistribution(System.out, 1000.0d);
     double completedMillis = (System.nanoTime() - start) / 1_000_000d;
     double rps = count / ((System.nanoTime() - start) / 1_000_000_000d);
-    System.out.println("test complete in " + completedMillis + "ms");
-    System.out.println("test rps " + rps);
-    System.out.println();
+    logger.info("test complete in {} ms", completedMillis);
+    logger.info("test rps {}", rps);
+  }
+
+  private static SimpleResponse toResponse(ObjectMapper mapper, InputStream response) {
+    try {
+      return mapper.readValue(response, SimpleResponse.class);
+    } catch (IOException e) {
+      logger.error(response, e);
+      throw new RuntimeException(e);
+    }
+  }
+
+  private static Mono<ByteBuf> newRequest(ObjectMapper mapper) {
+    SimpleRequest request = new SimpleRequest();
+    request.setMessage("hello");
+    try {
+      return Mono.just(Unpooled.wrappedBuffer(mapper.writeValueAsBytes(request)));
+    } catch (JsonProcessingException e) {
+      return Mono.error(e);
+    }
   }
 }
