+++ rsocket-rpc/src/main/proto/io/netifi/thunderdome/rsocketrpc/simpleservice.proto
